pyproject.toml
[project]
name = "bss"
version = "0.1.0"
description = "Blind Source Separation algorithms for audio processing"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "soundfile>=0.12.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
run-bss = "run_bss:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bss"]

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py"]
